# server/services/swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: SGL API
  description: API para el Sistema de Gestión de Laboratorios (SGL). Este sistema está diseñado para facilitar la gestión de equipos, usuarios, reservas, y mantenimiento de laboratorios.
  version: 1.0.0
  contact:
    name: Angelo G
    email: angelogarcia2021@@gmail.com
    url: https://wa.me/593982718318
servers:
- url: http://localhost:3000/api
  description: API local
paths:
...
  /user/all:
    get:
      summary: "Obtener todos los usuarios"
      description: |
        Esta ruta permite obtener una lista de todos los usuarios del sistema. 
        Se pueden aplicar filtros opcionales como nombre completo (`full_name`) y cédula de identificación (`identification_card`).
        Los resultados están paginados.

        **Parámetros de consulta:**
        - `page`: Número de página (por defecto: 1).
        - `limit`: Número de usuarios por página (por defecto: 10).
        - `full_name`: Filtro por nombre completo (opcional).
        - `identification_card`: Filtro por cédula de identificación (opcional).
      operationId: "getAllUsers"
      tags:
      - "Usuarios"
      parameters:
      - name: page
        in: query
        description: "Número de página (por defecto: 1)"
        required: false
        schema:
          type: integer
          example: 1
      - name: limit
        in: query
        description: "Número de registros por página (por defecto: 10)"
        required: false
        schema:
          type: integer
          example: 10
      - name: full_name
        in: query
        description: "Filtro opcional por nombre completo"
        required: false
        schema:
          type: string
          example: "Juan Pérez"
      - name: identification_card
        in: query
        description: "Filtro opcional por cédula de identificación"
        required: false
        schema:
          type: string
          example: "123456789"
      responses:
        "200":
          description: "Usuarios obtenidos exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Usuarios obtenidos exitosamente."
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id_user:
                            type: string
                            example: "uuid"
                          active:
                            type: boolean
                            example: true
                          full_name:
                            type: string
                            example: "string"
                          email:
                            type: string
                            example: "string"
                          phone:
                            type: string
                            example: "string"
                          identification_card:
                            type: string
                            example: "string"
                          code:
                            type: string
                            example: "string"
                          createdAt:
                            type: string
                            format: date-time
                            example: "date ISO"
                          updatedAt:
                            type: string
                            format: date-time
                            example: "date ISO"
                          user_roles_intermediate:
                            type: object
                            properties:
                              id_user_role_intermediate:
                                type: string
                                example: "uuid"
                              id_user_fk:
                                type: string
                                example: "uuid"
                              user_roles:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    id_user_roles:
                                      type: string
                                      example: "uuid"
                                    id_rol_fk:
                                      type: string
                                      example: "uuid"
                                    role:
                                      type: object
                                      properties:
                                        id_rol:
                                          type: string
                                          example: "uuid"
                                        type_rol:
                                          type: string
                                          example: "string"
        "404":
          description: Usuarios no encontrados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: Usuarios no encontrados.
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: ¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo.
  /user/get-by-id/{id}:
    get:
      summary: "Obtener usuario por ID"
      description: |
        Esta ruta permite obtener los detalles de un usuario específico mediante su identificador único (ID).

        **Funcionamiento:**
        - El cliente envía el ID del usuario como parámetro en la URL.
        - El servidor verifica si el usuario existe y devuelve sus datos.
        - Los resultados se almacenan temporalmente en caché para optimizar las siguientes solicitudes.

      operationId: "getUserById"
      tags:
      - "Usuarios"
      parameters:
      - name: UUID
        in: path
        description: "UUID del usuario"
        required: true
        schema:
          type: integer
          example: 1
      responses:
        "200":
          description: "Usuario obtenido exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Usuario obtenido exitosamente."
                  data:
                    type: object
                    properties:
                      user:
                        type: object
                        properties:
                          id_user:
                            type: string
                            example: "2e5a412d-bcd0-4484-acd0-dda02492d2c6"
                          active:
                            type: boolean
                            example: true
                          full_name:
                            type: string
                            example: "Administrador Acceso"
                          email:
                            type: string
                            example: "acceso@gmail.com"
                          phone:
                            type: string
                            example: "0000000009"
                          identification_card:
                            type: string
                            example: "0000000009"
                          code:
                            type: string
                            example: "AA"
                          createdAt:
                            type: string
                            format: date-time
                            example: "2024-12-10T11:02:23.426Z"
                          updatedAt:
                            type: string
                            format: date-time
                            example: "2024-12-10T11:02:23.426Z"
                          user_roles_intermediate:
                            type: object
                            properties:
                              id_user_role_intermediate:
                                type: string
                                example: "cd05c88c-becc-45c3-8747-31cb5472d285"
                              id_user_fk:
                                type: string
                                example: "2e5a412d-bcd0-4484-acd0-dda02492d2c6"
                              user_roles:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    id_user_roles:
                                      type: string
                                      example: "a4bac6a0-e419-4754-8342-52ca701b1339"
                                    id_rol_fk:
                                      type: string
                                      example: "ac5c8ace-0b24-4242-a5d4-d05acacd55d7"
                                    role:
                                      type: object
                                      properties:
                                        id_rol:
                                          type: string
                                          example: "ac5c8ace-0b24-4242-a5d4-d05acacd55d7"
                                        type_rol:
                                          type: string
                                          example: "access_manager"
        "404":
          description: Usuarios no encontrados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: Usuarios no encontrados.
        "500":
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: ¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo.
  /user/create:
    post:
      summary: "Crear nuevo usuario"
      description: |
        Esta ruta permite crear un nuevo usuario en el sistema.

        **Funcionamiento:**
        1. Se verifica si los campos obligatorios están presentes y son válidos.
        2. Se valida que no existan duplicados para campos únicos como correo electrónico, teléfono y número de cédula.
        3. Se crea un nuevo usuario y se genera un hash para la contraseña.
        4. Un correo de confirmación se envía tras la creación del usuario.

        **Códigos de estado:**
        - **201**: Usuario creado exitosamente.
        - **400**: Los datos enviados no son válidos o faltan campos obligatorios.
        - **409**: Ya existe un usuario con el mismo correo, teléfono o número de cédula.
        - **500**: Error en el servidor al procesar la solicitud.
      operationId: "createUser"
      tags:
      - "Usuarios"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Juan Pérez"
                email:
                  type: string
                  example: "juanperez@gmail.com"
                phone:
                  type: string
                  example: "1234567890"
                identification_card:
                  type: string
                  example: "V-12345678"
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "Juan Pérez"
                email:
                  type: string
                  example: "juanperez@gmail.com"
                phone:
                  type: string
                  example: "1234567890"
                identification_card:
                  type: string
                  example: "V-12345678"
      responses:
        "201":
          description: "Usuario creado exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "Usuario creado exitosamente."
                  data:
                    type: object
                    properties:
                      id_user:
                        type: string
                        example: uuid"
                      full_name:
                        type: string
                        example: "string"
                      email:
                        type: string
                        example: "string"
                      phone:
                        type: string
                        example: "string"
                      identification_card:
                        type: string
                        example: "string"
                      createdAt:
                        type: string
                        format: date-time
                        example: "date ISO"
                      updatedAt:
                        type: string
                        format: date-time
                        example: "date ISO"
        "400":
          description: "Solicitud incorrecta, datos faltantes o inválidos."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Ya existe un usuario con este correo."
        "404":
          description: "No se encontraron registros de usuario."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Usuarios no encontrados."
        "500":
          description: "Error interno del servidor."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error. Por favor, inténtalo de nuevo."
...
  /user/update-password:
    put:
      summary: "Actualizar la contraseña del usuario"
      description: |
        Esta ruta permite actualizar la contraseña de un usuario en el sistema. El usuario debe proporcionar su contraseña actual, la nueva contraseña y su confirmación. Si la contraseña actual es correcta y la nueva contraseña y confirmación coinciden, la contraseña será actualizada.

        **Flujo de actualización de contraseña:**
        1. El cliente envía la contraseña actual, la nueva contraseña y su confirmación.
        2. El servidor valida que la contraseña actual coincida con la almacenada en la base de datos.
        3. Se valida que la nueva contraseña y la confirmación coincidan.
        4. Si las validaciones son exitosas, el servidor genera un nuevo hash para la nueva contraseña.
        5. El servidor actualiza la contraseña en la base de datos.
        6. El servidor devuelve una respuesta de éxito indicando que la contraseña ha sido actualizada.

        **Códigos de estado:**
        - **200**: Contraseña actualizada exitosamente.
        - **400**: La nueva contraseña no coincide con la confirmación.
        - **401**: La contraseña actual es incorrecta.
        - **500**: Error en el servidor al procesar la solicitud de actualización de la contraseña.
      operationId: "updatePassword"
      tags:
      - "Usuarios"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: "password123"
                newPassword:
                  type: string
                  example: "newPassword456"
                confirmPassword:
                  type: string
                  example: "newPassword456"
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  example: "password123"
                newPassword:
                  type: string
                  example: "newPassword456"
                confirmPassword:
                  type: string
                  example: "newPassword456"
      responses:
        "200":
          description: "Contraseña actualizada exitosamente."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Contraseña actualizada exitosamente."
        "400":
          description: "La nueva contraseña no coincide con la confirmación."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "La nueva contraseña no coincide con la confirmación."
        "401":
          description: "Contraseña actual inválida."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: "Contraseña actual inválida."
        "500":
          description: "Error interno del servidor al procesar la solicitud de actualización de contraseña."
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "error"
                  code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "¡Ops! Ha ocurrido un error al actualizar la contraseña. Por favor, inténtalo de nuevo."
...